+++ contrib/opencensus-ext-azure/tests/test_transport_mixin.py
@@ -20,6 +20,8 @@ import unittest
 import mock
 import requests
 
+from azure.core.exceptions import ClientAuthenticationError
+from azure.identity._exceptions import CredentialUnavailableError
 from opencensus.ext.azure.common import Options
 from opencensus.ext.azure.common.storage import LocalFileStorage
 from opencensus.ext.azure.common.transport import TransportMixin
@@ -68,6 +70,39 @@ class TestTransportMixin(unittest.TestCase):
             self.assertIsNone(mixin.storage.get())
             self.assertEqual(len(os.listdir(mixin.storage.path)), 1)
 
+    def test_transmission_pre_req_exception(self):
+        mixin = TransportMixin()
+        mixin.options = Options()
+        with LocalFileStorage(os.path.join(TEST_FOLDER, self.id())) as stor:
+            mixin.storage = stor
+            mixin.storage.put([1, 2, 3])
+            with mock.patch('requests.post', throw(requests.RequestException)):
+                mixin._transmit_from_storage()
+            self.assertIsNone(mixin.storage.get())
+            self.assertEqual(len(os.listdir(mixin.storage.path)), 1)
+
+    def test_transmission_cred_exception(self):
+        mixin = TransportMixin()
+        mixin.options = Options()
+        with LocalFileStorage(os.path.join(TEST_FOLDER, self.id())) as stor:
+            mixin.storage = stor
+            mixin.storage.put([1, 2, 3])
+            with mock.patch('requests.post', throw(CredentialUnavailableError)):
+                mixin._transmit_from_storage()
+            self.assertIsNone(mixin.storage.get())
+            self.assertEqual(len(os.listdir(mixin.storage.path)), 0)
+
+    def test_transmission_client_exception(self):
+        mixin = TransportMixin()
+        mixin.options = Options()
+        with LocalFileStorage(os.path.join(TEST_FOLDER, self.id())) as stor:
+            mixin.storage = stor
+            mixin.storage.put([1, 2, 3])
+            with mock.patch('requests.post', throw(ClientAuthenticationError)):
+                mixin._transmit_from_storage()
+            self.assertIsNone(mixin.storage.get())
+            self.assertEqual(len(os.listdir(mixin.storage.path)), 1)
+
     def test_transmission_pre_exception(self):
         mixin = TransportMixin()
         mixin.options = Options()
@@ -77,7 +112,7 @@ class TestTransportMixin(unittest.TestCase):
             with mock.patch('requests.post', throw(Exception)):
                 mixin._transmit_from_storage()
             self.assertIsNone(mixin.storage.get())
-            self.assertEqual(len(os.listdir(mixin.storage.path)), 1)
+            self.assertEqual(len(os.listdir(mixin.storage.path)), 0)
 
     @mock.patch('requests.post', return_value=mock.Mock())
     def test_transmission_lease_failure(self, requests_mock):
@@ -119,6 +154,35 @@ class TestTransportMixin(unittest.TestCase):
                 mixin._transmit_from_storage()
             self.assertIsNone(mixin.storage.get())
             self.assertEqual(len(os.listdir(mixin.storage.path)), 0)
+
+    def test_transmission_auth(self):
+        mixin = TransportMixin()
+        mixin.options = Options()
+        credential = mock.Mock()
+        mixin.options.credential = credential
+        token_mock = mock.Mock()
+        token_mock.token = "test_token"
+        credential.get_token.return_value = token_mock
+        url = 'https://eastus-1.in.applicationinsights.azure.com//v2/track'
+        data = '[1, 2, 3]'
+        headers = {
+            'Accept': 'application/json',
+            'Content-Type': 'application/json; charset=utf-8',
+            'Authorization': 'Bearer test_token',
+        }
+        with LocalFileStorage(os.path.join(TEST_FOLDER, self.id())) as stor:
+            mixin.storage = stor
+            mixin.storage.put([1, 2, 3])
+            with mock.patch('requests.post') as post:
+                post.return_value = MockResponse(200, 'unknown')
+                mixin._transmit_from_storage()
+                post.assert_called_with(
+                    url=url, data=data, headers=headers, timeout=10.0, proxies={})
+            credential.get_token.assert_called_with(
+                "https://monitor.azure.com/.default")
+            self.assertIsNone(mixin.storage.get())
+            self.assertEqual(len(os.listdir(mixin.storage.path)), 0)
+            credential.get_token.assert_called_once()
 
     def test_transmission_206(self):
         mixin = TransportMixin()
@@ -201,16 +265,16 @@ class TestTransportMixin(unittest.TestCase):
             self.assertIsNone(mixin.storage.get())
             self.assertEqual(len(os.listdir(mixin.storage.path)), 0)
 
-    def test_transmission_400(self):
+    def test_transmission_401(self):
         mixin = TransportMixin()
         mixin.options = Options()
         with LocalFileStorage(os.path.join(TEST_FOLDER, self.id())) as stor:
             mixin.storage = stor
             mixin.storage.put([1, 2, 3])
             with mock.patch('requests.post') as post:
-                post.return_value = MockResponse(400, '{}')
+                post.return_value = MockResponse(401, '{}')
                 mixin._transmit_from_storage()
-            self.assertEqual(len(os.listdir(mixin.storage.path)), 0)
+            self.assertEqual(len(os.listdir(mixin.storage.path)), 1)
 
     def test_transmission_500(self):
         mixin = TransportMixin()
@@ -223,3 +287,14 @@ class TestTransportMixin(unittest.TestCase):
                 mixin._transmit_from_storage()
             self.assertIsNone(mixin.storage.get())
             self.assertEqual(len(os.listdir(mixin.storage.path)), 1)
+
+    def test_transmission_400(self):
+        mixin = TransportMixin()
+        mixin.options = Options()
+        with LocalFileStorage(os.path.join(TEST_FOLDER, self.id())) as stor:
+            mixin.storage = stor
+            mixin.storage.put([1, 2, 3])
+            with mock.patch('requests.post') as post:
+                post.return_value = MockResponse(400, '{}')
+                mixin._transmit_from_storage()
+            self.assertEqual(len(os.listdir(mixin.storage.path)), 0)